requirements.txt
asgiref==3.8.1
Django==5.2.1
djangorestframework==3.16.0
mysqlclient==2.2.7
python-dotenv==1.1.0
sqlparse==0.5.3
gunicorn==21.2.0
django-filter>=23.0
pytest
pytest-django
factory_boy
coverage